Remove pattern and change max year value
Previously there was a "pattern" attribute on each of the html input fields.

This contravenes the W3C standards for fields with type="number", and so has been removed.

Also, the maximum value for the year field was set to the value of the current year, which would
entry of dates in the future.  As this gem is intended for generalised use, the maximum value has
been set to 2100, and the client should validate as necessary.