Add option to control behaviour of multiple error messages
If invalid values are entered on a gov_uk_date_field,  e.g. 35/16/2016,
this gem will add an Invalid date error message to the error hash for the field,
but the underlying model may also have added a message, such as "Cannot be blank"
(because the date field will be nil in this case).

This PR adds a new option to the acts_as_gov_uk_date_field method to allow the
client to control the behaviour in this case, either displaying both error messages,
dropping the gem-generated error message or dropping the model-generated error message.