Change how we access attributes in the ActiveRecord object
Change the write accessor defined in this library to rely on 'super' instead of
the writer method 'object[attr_name]=', with the implication that writing the
date now has to happen in the write accessor.

While implementing gov uk date fields with jsonb attributes, we were running
into an issue with dates that act as gov uk dates. The dates were not being
saved properly to the JSON. On investigation we found that this was because the
gov_uk_date_fields library was writing attributes with 'object[attr_name]=' and
reading them with 'object[attr_name]' instead of 'object.attr_name=' and
'object.attr_name'. The jsonb_accessor (and incidentally other libraries like
papertrail) override 'object.attr_name=' and can be re-used in a writer method
by calling 'super'.