Remove after_initialize duplication in acts_as_gov_uk_date
The same method was defined twice, with duplicate 'after_initialize' calls, but
only one was being used (the one defined with 'define_method') and the other was
missing a bit of code to make it work. Because we are adding code to models,
including a module with our methods is preferable so we've left that version.