updates markup and styling of errors
This commit updates the markup that this gem produces so that its closer
inline with the markup that GOVUK Elements uses

https://govuk-elements.herokuapp.com/form-elements/example-date/

It also fixes an issue with error styling which was broken as a result
of GDS changing the css class name used to style error messages. The
class name changes from `error` to `form-group-error` also each input
field also receives a `form-control-error`

https://github.com/alphagov/govuk_elements/releases/tag/v3.0.0
https://github.com/alphagov/govuk_elements/pull/405

As this commit introduces new markup and potentially breaking changes
for any services relying on the markup for styling or js interaction
may want to upgrade with caution.

In addition to this update if your project is using GOVUK Elements < 3.0
then you will also need to update your GOVUK Elements version