Generate group id (fix error summary linking)
The gem already supported a way to specify the `id` of the options passed to the constructor, but with this PR the id, if not provided, will be autogenerated based on the form object and the attribute of the date field.

This solves the issue where the errors in the summary where not linking correctly (because the id in the form element was blank or incorrect).

It is backwards compatible as much as possible. If you were already specifying an id, this will take precedence over the autogenerated one. If no id is passed, then an autogenerated one will be added to the form element.