Bump version from 3.0.0 to 3.1.0
And fix most of the security alerts in dependencies, but in a conservative way as this gem is suppose to be used in the context of a full app that will be in control of actually specifying the versions of the dependencies they want to use.

I don't want to "force" specific versions in this gem.