Allow alternative i18n locales
There are situations where, even if the attribute is the same, the legend and/or hint must be dynamic to further customise the view.

This is usually the case with partials where reusing a form element in multiple places through partials has the side effect of the attribute being always the same and thus the locale also the same.

To improve this we are adding a new option to pass optionally, named `i18n_attribute` that will be the key to lookup in hint and legend locales.

The "real" attribute will still take precedence in all other places or, if the `i18n_attribute` locale is not found, to allow to DRY locales.

This change is backward compatible.