Change input number to text, and add inputmode
Following report of issues with assistive technologies and compatibility with input `type="number"` [0], this PR will make the recommended changes and convert these input to be of `type="text"` instead, also adding the extra attribute `inputmode="numeric"`.

Although this is a very minor change, the version of the gem has been bumped to 4.1.0 as it might be a breaking change for some code, previously relying on the type number.

[0] https://github.com/alphagov/govuk-frontend/issues/1449